SMM/requirements_video.txt
moviepy>=1.0.3
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
imageio-ffmpeg>=0.4.9

